[TASK] Subscribe to rippled manifests subscriptions
