+++ test/services/SubscriptionService.js
@@ -1,7 +1,7 @@
 import 'sails-test-helper'
 import assert from 'assert'
 
-describe('ValidationSubscriptionService', () => {
+describe('SubscriptionService', () => {
 
   before(async() => {
     await database.Validations.truncate()
@@ -15,7 +15,7 @@ describe('ValidationSubscriptionService', () => {
 
     it('should return a list of rippled nodes', async() => {
 
-      const rippleds = await ValidationSubscriptionService.getRippleds(process.env.PEERS_API_URL)
+      const rippleds = await SubscriptionService.getRippleds(process.env.PEERS_API_URL)
       expect(rippleds).to.be.instanceof(Array)
       expect(rippleds[0].version).to.exist
       expect(rippleds[0].public_key).to.exist
@@ -28,17 +28,16 @@ describe('ValidationSubscriptionService', () => {
 
       const rippleds = [{
         ipp: 's1.ripple.com:51235',
-        version: 'rippled-0.30.0',
-        public_key: 'n94szsuScMEjP4Nu7c3mndcyjWJELepZcASEkDMcDSwXhHH6Sn37'
+        version: 'rippled-0.30.0'
       }]
-      const connections = await ValidationSubscriptionService.subscribeToRippleds(rippleds)
+      const connections = await SubscriptionService.subscribeToRippleds(rippleds)
       // Wait for subscription to start
       setTimeout(() => {
         expect(connections).to.be.an('object')
-        expect(_.size(connections)).to.equal(1)
+        expect(_.size(connections)).to.equal(2)
         _.forEach(connections, function(connection, ip) {
-          expect(connection.public_key).to.equal(rippleds[0].public_key)
-          expect(connection.ws).to.exist
+          expect(['ws://s1.ripple.com:51233', 'wss://s.altnet.rippletest.net:51233']).to.include(ip)
+          expect(connection).to.exist
         });
         done()
       }, 1000)
